+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased] - yyyy-mm-dd
+
+### Fixed
+
+- Can not define more then one unaffiliated corporation as "main" (#63)
+
 ## [1.20.2] - 2022-01-08
 
 ### Fixed